backend/src/ai_agents/requirements.txt
flask
python-dotenv
groq
gunicorn
flask_cors